+++ std/commitments/kzg/verifier_test.go
@@ -3,31 +3,17 @@ package kzg
 import (
 	"crypto/rand"
 	"fmt"
+	"math/big"
 	"testing"
 
 	"github.com/consensys/gnark-crypto/ecc"
-	bls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377"
-	fr_bls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377/fr"
-	kzg_bls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377/kzg"
-	bls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381"
-	fr_bls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381/fr"
-	kzg_bls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381/kzg"
-	bls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315"
-	fr_bls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315/fr"
-	kzg_bls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315/kzg"
-	"github.com/consensys/gnark-crypto/ecc/bn254"
 	fr_bn254 "github.com/consensys/gnark-crypto/ecc/bn254/fr"
 	kzg_bn254 "github.com/consensys/gnark-crypto/ecc/bn254/kzg"
-	bw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761"
-	fr_bw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761/fr"
-	kzg_bw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761/kzg"
+	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/frontend"
-	"github.com/consensys/gnark/std/algebra"
-	"github.com/consensys/gnark/std/algebra/emulated/sw_bls12381"
 	"github.com/consensys/gnark/std/algebra/emulated/sw_bn254"
-	"github.com/consensys/gnark/std/algebra/emulated/sw_bw6761"
-	"github.com/consensys/gnark/std/algebra/native/sw_bls12377"
-	"github.com/consensys/gnark/std/algebra/native/sw_bls24315"
+	"github.com/consensys/gnark/std/math/emulated"
+	"github.com/consensys/gnark/std/recursion"
 	"github.com/consensys/gnark/test"
 )
 
@@ -36,29 +22,29 @@ const (
 	polynomialSize = 100
 )
 
-type KZGVerificationCircuit[S algebra.ScalarT, G1El algebra.G1ElementT, G2El algebra.G2ElementT, GTEl algebra.GtElementT] struct {
-	VerifyingKey[G2El]
-	Commitment[G1El]
-	OpeningProof[S, G1El]
+//--------------------------------------------------------
+// Single opening single point
+
+type KZGVerificationCircuit[S emulated.FieldParams, G1El, G2El, GTEl any] struct {
+	Vk     VerifyingKey[G1El, G2El]
+	Digest Commitment[G1El]
+	Proof  OpeningProof[S, G1El]
+	Point  emulated.Element[S]
 }
 
 func (c *KZGVerificationCircuit[S, G1El, G2El, GTEl]) Define(api frontend.API) error {
-	curve, err := algebra.GetCurve[S, G1El](api)
-	if err != nil {
-		return fmt.Errorf("get curve: %w", err)
-	}
-	pairing, err := algebra.GetPairing[G1El, G2El, GTEl](api)
+	verifier, err := NewVerifier[S, G1El, G2El, GTEl](api)
 	if err != nil {
 		return fmt.Errorf("get pairing: %w", err)
 	}
-	verifier := NewVerifier(c.VerifyingKey, curve, pairing)
-	if err := verifier.AssertProof(c.Commitment, c.OpeningProof); err != nil {
+	if err := verifier.CheckOpeningProof(c.Digest, c.Proof, c.Point, c.Vk); err != nil {
 		return fmt.Errorf("assert proof: %w", err)
 	}
 	return nil
 }
 
 func TestKZGVerificationEmulated(t *testing.T) {
+
 	assert := test.NewAssert(t)
 
 	alpha, err := rand.Int(rand.Reader, ecc.BN254.ScalarField())
@@ -85,329 +71,490 @@ func TestKZGVerificationEmulated(t *testing.T) {
 
 	wCmt, err := ValueOfCommitment[sw_bn254.G1Affine](com)
 	assert.NoError(err)
-	wProof, err := ValueOfOpeningProof[sw_bn254.Scalar, sw_bn254.G1Affine](point, proof)
+	wProof, err := ValueOfOpeningProof[emulated.BN254Fr, sw_bn254.G1Affine](proof)
 	assert.NoError(err)
-	wVk, err := ValueOfVerifyingKey[sw_bn254.G2Affine](srs.Vk)
+	wVk, err := ValueOfVerifyingKey[sw_bn254.G1Affine, sw_bn254.G2Affine](srs.Vk)
 	assert.NoError(err)
 
-	assignment := KZGVerificationCircuit[sw_bn254.Scalar, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{
-		VerifyingKey: wVk,
-		Commitment:   wCmt,
-		OpeningProof: wProof,
+	// wPoint, err := ValueOfScalar[emulated.BN254Fr](point)
+	wPoint, err := ValueOfScalar[emulated.BN254Fr](point)
+	assert.NoError(err)
+
+	assignment := KZGVerificationCircuit[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{
+		Vk:     wVk,
+		Digest: wCmt,
+		Proof:  wProof,
+		Point:  wPoint,
 	}
-	assert.CheckCircuit(&KZGVerificationCircuit[sw_bn254.Scalar, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{}, test.WithValidAssignment(&assignment))
+	assert.CheckCircuit(&KZGVerificationCircuit[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{}, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BLS12_381), test.WithBackends(backend.PLONK))
 }
 
-func TestKZGVerificationEmulated2(t *testing.T) {
-	assert := test.NewAssert(t)
+//--------------------------------------------------------
+// Slice proof
 
-	alpha, err := rand.Int(rand.Reader, ecc.BLS12_381.ScalarField())
-	assert.NoError(err)
-	srs, err := kzg_bls12381.NewSRS(kzgSize, alpha)
-	assert.NoError(err)
+// type SliceTest[S emulated.FieldParams, G1El any] struct {
+// 	Proof BatchOpeningProof[S, G1El]
+// }
 
-	f := make([]fr_bls12381.Element, polynomialSize)
-	for i := range f {
-		f[i].SetRandom()
+// func (circuit *SliceTest[S, G1El]) Define(api frontend.API) error {
+
+// 	fmt.Println(len(circuit.Proof.ClaimedValues))
+// 	return nil
+// }
+
+// func TestSlices(t *testing.T) {
+
+// 	assert := test.NewAssert(t)
+
+// 	// prepare test data
+// 	alpha, err := rand.Int(rand.Reader, ecc.BN254.ScalarField())
+// 	assert.NoError(err)
+// 	srs, err := kzg_bn254.NewSRS(kzgSize, alpha)
+// 	assert.NoError(err)
+
+// 	var polynomials [10][]fr_bn254.Element
+// 	var coms [10]kzg_bn254.Digest
+// 	for i := 0; i < 10; i++ {
+// 		polynomials[i] = make([]fr_bn254.Element, polynomialSize)
+// 		for j := 0; j < polynomialSize; j++ {
+// 			polynomials[i][j].SetRandom()
+// 		}
+// 		coms[i], err = kzg_bn254.Commit(polynomials[i], srs.Pk)
+// 		assert.NoError(err)
+// 	}
+
+// 	var point fr_bn254.Element
+// 	point.SetRandom()
+// 	var target big.Int
+// 	target.SetUint64(1)
+// 	nbBits := ecc.BLS12_381.ScalarField().BitLen()
+// 	nn := ((nbBits+7)/8)*8 - 8
+// 	target.Lsh(&target, uint(nn))
+// 	h, err := recursion.NewShort(ecc.BLS12_381.ScalarField(), &target)
+// 	assert.NoError(err)
+
+// 	batchOpeningProof, err := kzg_bn254.BatchOpenSinglePoint(polynomials[:], coms[:], point, h, srs.Pk)
+// 	assert.NoError(err)
+
+// 	_, _, err = kzg_bn254.FoldProof(coms[:], &batchOpeningProof, point, h)
+// 	assert.NoError(err)
+
+// 	// circuit
+// 	var circuit SliceTest[emulated.BN254Fr, sw_bn254.G1Affine]
+// 	circuit.Proof.ClaimedValues = make([]emulated.Element[emulated.BN254Fr], 10)
+
+// 	var assignment SliceTest[emulated.BN254Fr, sw_bn254.G1Affine]
+// 	// assignment.Proof.ClaimedValues = make([]emulated.Element[emulated.BN254Fr], 10)
+// 	assignment.Proof, err = ValueOfBatchOpeningProof[emulated.BN254Fr, sw_bn254.G1Affine](batchOpeningProof)
+// 	assert.NoError(err)
+
+// 	assert.CheckCircuit(&circuit, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BLS12_381), test.WithBackends(backend.PLONK))
+
+// }
+
+//--------------------------------------------------------
+// Fold proof
+
+type FoldProofTest[S emulated.FieldParams, G1El, G2El, GTEl any] struct {
+	Point                emulated.Element[S]
+	Digests              [10]Commitment[G1El]
+	BatchOpeningProof    BatchOpeningProof[S, G1El]
+	ExpectedFoldedProof  OpeningProof[S, G1El]
+	ExpectedFoldedDigest Commitment[G1El]
+}
+
+func (c *FoldProofTest[S, G1El, G2El, GTEl]) Define(api frontend.API) error {
+
+	verifier, err := NewVerifier[S, G1El, G2El, GTEl](api)
+	if err != nil {
+		return fmt.Errorf("get pairing: %w", err)
 	}
 
-	com, err := kzg_bls12381.Commit(f, srs.Pk)
+	// pick a number on byte shorter than the modulus size
+	var target big.Int
+	target.SetUint64(1)
+	nbBits := api.Compiler().Field().BitLen()
+	nn := ((nbBits+7)/8)*8 - 8
+	target.Lsh(&target, uint(nn))
+
+	// create the wrapped hash function
+	whSnark, err := recursion.NewHash(api, &target, true)
+	if err != nil {
+		return err
+	}
+
+	// op, com, err := verifier.FoldProof(c.Digests[:], c.BatchOpeningProof, c.Point, whSnark)
+	foldedProof, foldedDigests, err := verifier.FoldProof(c.Digests[:], c.BatchOpeningProof, c.Point, whSnark)
+	if err != nil {
+		return err
+	}
+
+	verifier.ec.AssertIsEqual(&foldedDigests.G1El, &c.ExpectedFoldedDigest.G1El)
+	verifier.ec.AssertIsEqual(&foldedProof.Quotient, &c.ExpectedFoldedProof.Quotient)
+	verifier.scalarApi.AssertIsEqual(&foldedProof.ClaimedValue, &c.ExpectedFoldedProof.ClaimedValue)
+
+	return nil
+}
+
+func TestFoldProof(t *testing.T) {
+
+	assert := test.NewAssert(t)
+
+	// prepare test data
+	alpha, err := rand.Int(rand.Reader, ecc.BN254.ScalarField())
+	assert.NoError(err)
+	srs, err := kzg_bn254.NewSRS(kzgSize, alpha)
 	assert.NoError(err)
 
-	var point fr_bls12381.Element
+	var polynomials [10][]fr_bn254.Element
+	var coms [10]kzg_bn254.Digest
+	for i := 0; i < 10; i++ {
+		polynomials[i] = make([]fr_bn254.Element, polynomialSize)
+		for j := 0; j < polynomialSize; j++ {
+			polynomials[i][j].SetRandom()
+		}
+		coms[i], err = kzg_bn254.Commit(polynomials[i], srs.Pk)
+		assert.NoError(err)
+	}
+
+	var point fr_bn254.Element
 	point.SetRandom()
-	proof, err := kzg_bls12381.Open(f, point, srs.Pk)
+	var target big.Int
+	target.SetUint64(1)
+	nbBits := ecc.BLS12_381.ScalarField().BitLen()
+	nn := ((nbBits+7)/8)*8 - 8
+	target.Lsh(&target, uint(nn))
+	h, err := recursion.NewShort(ecc.BLS12_381.ScalarField(), &target)
 	assert.NoError(err)
 
-	if err = kzg_bls12381.Verify(&com, &proof, point, srs.Vk); err != nil {
-		t.Fatal("verify proof", err)
+	batchOpeningProof, err := kzg_bn254.BatchOpenSinglePoint(polynomials[:], coms[:], point, h, srs.Pk)
+	assert.NoError(err)
+
+	foldedProofs, foldedDigest, err := kzg_bn254.FoldProof(coms[:], &batchOpeningProof, point, h)
+	assert.NoError(err)
+
+	// prepare witness
+	wPoint, err := ValueOfScalar[emulated.BN254Fr](point)
+	assert.NoError(err)
+	var wDigests [10]Commitment[sw_bn254.G1Affine]
+	for i := 0; i < 10; i++ {
+		wDigests[i], err = ValueOfCommitment[sw_bn254.G1Affine](coms[i])
+		assert.NoError(err)
+	}
+	wBatchOpeningProof, err := ValueOfBatchOpeningProof[emulated.BN254Fr, sw_bn254.G1Affine](batchOpeningProof)
+	assert.NoError(err)
+	wExpectedFoldedProof, err := ValueOfOpeningProof[emulated.BN254Fr, sw_bn254.G1Affine](foldedProofs)
+	assert.NoError(err)
+	wExpectedFoldedDigest, err := ValueOfCommitment[sw_bn254.G1Affine](foldedDigest)
+	assert.NoError(err)
+
+	assignment := FoldProofTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{
+		Point:                wPoint,
+		Digests:              wDigests,
+		BatchOpeningProof:    wBatchOpeningProof,
+		ExpectedFoldedProof:  wExpectedFoldedProof,
+		ExpectedFoldedDigest: wExpectedFoldedDigest,
 	}
 
-	wCmt, err := ValueOfCommitment[sw_bls12381.G1Affine](com)
-	assert.NoError(err)
-	wProof, err := ValueOfOpeningProof[sw_bls12381.Scalar, sw_bls12381.G1Affine](point, proof)
-	assert.NoError(err)
-	wVk, err := ValueOfVerifyingKey[sw_bls12381.G2Affine](srs.Vk)
-	assert.NoError(err)
+	var circuit FoldProofTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]
+	circuit.BatchOpeningProof.ClaimedValues = make([]emulated.Element[emulated.BN254Fr], 10)
+	assert.CheckCircuit(&circuit, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BLS12_381), test.WithBackends(backend.PLONK))
 
-	assignment := KZGVerificationCircuit[sw_bls12381.Scalar, sw_bls12381.G1Affine, sw_bls12381.G2Affine, sw_bls12381.GTEl]{
-		VerifyingKey: wVk,
-		Commitment:   wCmt,
-		OpeningProof: wProof,
-	}
-	assert.CheckCircuit(&KZGVerificationCircuit[sw_bls12381.Scalar, sw_bls12381.G1Affine, sw_bls12381.G2Affine, sw_bls12381.GTEl]{}, test.WithValidAssignment(&assignment))
 }
 
-func TestKZGVerificationEmulated3(t *testing.T) {
-	assert := test.NewAssert(t)
+//--------------------------------------------------------
+// Batch verify single point
 
-	alpha, err := rand.Int(rand.Reader, ecc.BW6_761.ScalarField())
-	assert.NoError(err)
-	srs, err := kzg_bw6761.NewSRS(kzgSize, alpha)
-	assert.NoError(err)
+type BatchVerifySinglePointTest[S emulated.FieldParams, G1El, G2El, GTEl any] struct {
+	Vk                VerifyingKey[G1El, G2El]
+	Point             emulated.Element[S]
+	Digests           [10]Commitment[G1El]
+	BatchOpeningProof BatchOpeningProof[S, G1El]
+}
 
-	f := make([]fr_bw6761.Element, polynomialSize)
-	for i := range f {
-		f[i].SetRandom()
+func (c *BatchVerifySinglePointTest[S, G1El, G2El, GTEl]) Define(api frontend.API) error {
+
+	verifier, err := NewVerifier[S, G1El, G2El, GTEl](api)
+	if err != nil {
+		return fmt.Errorf("get pairing: %w", err)
 	}
 
-	com, err := kzg_bw6761.Commit(f, srs.Pk)
+	// pick a number on byte shorter than the modulus size
+	var target big.Int
+	target.SetUint64(1)
+	nbBits := api.Compiler().Field().BitLen()
+	nn := ((nbBits+7)/8)*8 - 8
+	target.Lsh(&target, uint(nn))
+
+	// create the wrapped hash function
+	whSnark, err := recursion.NewHash(api, &target, true)
+	if err != nil {
+		return err
+	}
+
+	// op, com, err := verifier.FoldProof(c.Digests[:], c.BatchOpeningProof, c.Point, whSnark)
+	verifier.BatchVerifySinglePoint(c.Digests[:], c.BatchOpeningProof, c.Point, whSnark, c.Vk)
+
+	return nil
+}
+
+func TestBatchVerifySinglePoint(t *testing.T) {
+
+	assert := test.NewAssert(t)
+
+	// prepare test data
+	alpha, err := rand.Int(rand.Reader, ecc.BN254.ScalarField())
+	assert.NoError(err)
+	srs, err := kzg_bn254.NewSRS(kzgSize, alpha)
 	assert.NoError(err)
 
-	var point fr_bw6761.Element
+	var polynomials [10][]fr_bn254.Element
+	var coms [10]kzg_bn254.Digest
+	for i := 0; i < 10; i++ {
+		polynomials[i] = make([]fr_bn254.Element, polynomialSize)
+		for j := 0; j < polynomialSize; j++ {
+			polynomials[i][j].SetRandom()
+		}
+		coms[i], err = kzg_bn254.Commit(polynomials[i], srs.Pk)
+		assert.NoError(err)
+	}
+
+	// random point at which the polynomials are evaluated
+	var point fr_bn254.Element
 	point.SetRandom()
-	proof, err := kzg_bw6761.Open(f, point, srs.Pk)
+
+	// build short hash, we pick a number one byte less than the snark field...
+	var target big.Int
+	target.SetUint64(1)
+	nbBits := ecc.BLS12_381.ScalarField().BitLen()
+	nn := ((nbBits+7)/8)*8 - 8
+	target.Lsh(&target, uint(nn))
+	h, err := recursion.NewShort(ecc.BLS12_381.ScalarField(), &target)
 	assert.NoError(err)
 
-	if err = kzg_bw6761.Verify(&com, &proof, point, srs.Vk); err != nil {
-		t.Fatal("verify proof", err)
+	batchOpeningProof, err := kzg_bn254.BatchOpenSinglePoint(polynomials[:], coms[:], point, h, srs.Pk)
+	assert.NoError(err)
+
+	err = kzg_bn254.BatchVerifySinglePoint(coms[:], &batchOpeningProof, point, h, srs.Vk)
+	assert.NoError(err)
+
+	// prepare witness
+	wVk, err := ValueOfVerifyingKey[sw_bn254.G1Affine, sw_bn254.G2Affine](srs.Vk)
+	assert.NoError(err)
+	wPoint, err := ValueOfScalar[emulated.BN254Fr](point)
+	assert.NoError(err)
+	var wDigests [10]Commitment[sw_bn254.G1Affine]
+	for i := 0; i < 10; i++ {
+		wDigests[i], err = ValueOfCommitment[sw_bn254.G1Affine](coms[i])
+		assert.NoError(err)
+	}
+	wBatchOpeningProof, err := ValueOfBatchOpeningProof[emulated.BN254Fr, sw_bn254.G1Affine](batchOpeningProof)
+	assert.NoError(err)
+
+	assignment := BatchVerifySinglePointTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{
+		Vk:                wVk,
+		Point:             wPoint,
+		Digests:           wDigests,
+		BatchOpeningProof: wBatchOpeningProof,
 	}
 
-	wCmt, err := ValueOfCommitment[sw_bw6761.G1Affine](com)
-	assert.NoError(err)
-	wProof, err := ValueOfOpeningProof[sw_bw6761.Scalar, sw_bw6761.G1Affine](point, proof)
-	assert.NoError(err)
-	wVk, err := ValueOfVerifyingKey[sw_bw6761.G2Affine](srs.Vk)
-	assert.NoError(err)
+	var circuit BatchVerifySinglePointTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]
+	circuit.BatchOpeningProof.ClaimedValues = make([]emulated.Element[emulated.BN254Fr], 10)
+	assert.CheckCircuit(&circuit, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BLS12_381), test.WithBackends(backend.PLONK))
 
-	assignment := KZGVerificationCircuit[sw_bw6761.Scalar, sw_bw6761.G1Affine, sw_bw6761.G2Affine, sw_bw6761.GTEl]{
-		VerifyingKey: wVk,
-		Commitment:   wCmt,
-		OpeningProof: wProof,
-	}
-	assert.CheckCircuit(&KZGVerificationCircuit[sw_bw6761.Scalar, sw_bw6761.G1Affine, sw_bw6761.G2Affine, sw_bw6761.GTEl]{}, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BN254))
 }
 
-func TestKZGVerificationTwoChain(t *testing.T) {
+//--------------------------------------------------------
+// Batch verify multi point
+
+// type BatchVerifyMultiPointsTest[S emulated.FieldParams, G1El, G2El, GTEl any] struct {
+// 	Vk      VerifyingKey[G1El, G2El]
+// 	Digests [5]Commitment[G1El]
+// 	Proofs  [5]OpeningProof[S, G1El]
+// 	Points  [5]emulated.Element[S]
+// }
+
+type BatchVerifyMultiPointsTest[S emulated.FieldParams, G1El, G2El, GTEl any] struct {
+	Vk      VerifyingKey[G1El, G2El]
+	Digests [2]Commitment[G1El]
+	Proofs  [2]OpeningProof[S, G1El]
+	Points  [2]emulated.Element[S]
+}
+
+func (circuit *BatchVerifyMultiPointsTest[S, G1El, G2El, GTEl]) Define(api frontend.API) error {
+
+	verifier, err := NewVerifier[S, G1El, G2El, GTEl](api)
+	if err != nil {
+		return fmt.Errorf("get pairing: %w", err)
+	}
+
+	verifier.BatchVerifyMultiPoints(circuit.Digests[:], circuit.Proofs[:], circuit.Points[:], circuit.Vk)
+
+	return nil
+}
+
+func TestBatchVerifyMultiPoints(t *testing.T) {
+
 	assert := test.NewAssert(t)
 
-	alpha, err := rand.Int(rand.Reader, ecc.BLS12_377.ScalarField())
+	// prepare test data
+	alpha, err := rand.Int(rand.Reader, ecc.BN254.ScalarField())
 	assert.NoError(err)
-	srs, err := kzg_bls12377.NewSRS(kzgSize, alpha)
+	srs, err := kzg_bn254.NewSRS(kzgSize, alpha)
 	assert.NoError(err)
 
-	f := make([]fr_bls12377.Element, polynomialSize)
-	for i := range f {
-		f[i].SetRandom()
+	var polynomials [2][]fr_bn254.Element
+	var coms [2]kzg_bn254.Digest
+	for i := 0; i < 2; i++ {
+		polynomials[i] = make([]fr_bn254.Element, polynomialSize)
+		for j := 0; j < polynomialSize; j++ {
+			polynomials[i][j].SetRandom()
+		}
+		coms[i], err = kzg_bn254.Commit(polynomials[i], srs.Pk)
+		assert.NoError(err)
 	}
 
-	com, err := kzg_bls12377.Commit(f, srs.Pk)
-	assert.NoError(err)
-
-	var point fr_bls12377.Element
-	point.SetRandom()
-	proof, err := kzg_bls12377.Open(f, point, srs.Pk)
-	assert.NoError(err)
-
-	if err = kzg_bls12377.Verify(&com, &proof, point, srs.Vk); err != nil {
-		t.Fatal("verify proof", err)
+	// random points at which the polynomials are evaluated
+	var points [2]fr_bn254.Element
+	for i := 0; i < 2; i++ {
+		points[i].SetRandom()
 	}
 
-	wCmt, err := ValueOfCommitment[sw_bls12377.G1Affine](com)
-	assert.NoError(err)
-	wProof, err := ValueOfOpeningProof[sw_bls12377.Scalar, sw_bls12377.G1Affine](point, proof)
-	assert.NoError(err)
-	wVk, err := ValueOfVerifyingKey[sw_bls12377.G2Affine](srs.Vk)
-	assert.NoError(err)
-
-	assignment := KZGVerificationCircuit[sw_bls12377.Scalar, sw_bls12377.G1Affine, sw_bls12377.G2Affine, sw_bls12377.GT]{
-		VerifyingKey: wVk,
-		Commitment:   wCmt,
-		OpeningProof: wProof,
+	// build opening proofs
+	var openingProofs [2]kzg_bn254.OpeningProof
+	for i := 0; i < 2; i++ {
+		openingProofs[i], err = kzg_bn254.Open(polynomials[i], points[i], srs.Pk)
+		assert.NoError(err)
 	}
 
-	assert.CheckCircuit(&KZGVerificationCircuit[sw_bls12377.Scalar, sw_bls12377.G1Affine, sw_bls12377.G2Affine, sw_bls12377.GT]{}, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BW6_761))
+	// check that the proofs are correct
+	err = kzg_bn254.BatchVerifyMultiPoints(coms[:], openingProofs[:], points[:], srs.Vk)
+	assert.NoError(err)
+
+	// prepare witness
+	wVk, err := ValueOfVerifyingKey[sw_bn254.G1Affine, sw_bn254.G2Affine](srs.Vk)
+	assert.NoError(err)
+	var wDigests [2]Commitment[sw_bn254.G1Affine]
+	var wPoints [2]emulated.Element[emulated.BN254Fr]
+	var wOpeningProofs [2]OpeningProof[emulated.BN254Fr, sw_bn254.G1Affine]
+	for i := 0; i < 2; i++ {
+		wPoints[i], err = ValueOfScalar[emulated.BN254Fr](points[i])
+		assert.NoError(err)
+		wDigests[i], err = ValueOfCommitment[sw_bn254.G1Affine](coms[i])
+		assert.NoError(err)
+		wOpeningProofs[i], err = ValueOfOpeningProof[emulated.BN254Fr, sw_bn254.G1Affine](openingProofs[i])
+		assert.NoError(err)
+	}
+
+	assignment := BatchVerifyMultiPointsTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{
+		Vk:      wVk,
+		Points:  wPoints,
+		Digests: wDigests,
+		Proofs:  wOpeningProofs,
+	}
+
+	var circuit BatchVerifyMultiPointsTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]
+	assert.CheckCircuit(&circuit, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BLS12_381), test.WithBackends(backend.PLONK))
+
 }
 
-func TestKZGVerificationTwoChain2(t *testing.T) {
-	assert := test.NewAssert(t)
+//--------------------------------------------------------
+// derive gamma
 
-	alpha, err := rand.Int(rand.Reader, ecc.BLS24_315.ScalarField())
-	assert.NoError(err)
-	srs, err := kzg_bls24315.NewSRS(kzgSize, alpha)
-	assert.NoError(err)
+// type DeriveGammaTest[S emulated.FieldParams, G1El, G2El, GTEl any] struct {
+// 	Point         emulated.Element[S]
+// 	Digests       [10]Commitment[G1El]
+// 	ClaimedValues [10]emulated.Element[S]
+// 	Gamma         emulated.Element[S]
+// }
 
-	f := make([]fr_bls24315.Element, polynomialSize)
-	for i := range f {
-		f[i].SetRandom()
-	}
+// func (circuit *DeriveGammaTest[S, G1El, G2El, GTEl]) Define(api frontend.API) error {
 
-	com, err := kzg_bls24315.Commit(f, srs.Pk)
-	assert.NoError(err)
+// 	verifier, err := NewVerifier[S, G1El, G2El, GTEl](api)
+// 	if err != nil {
+// 		return fmt.Errorf("get pairing: %w", err)
+// 	}
 
-	var point fr_bls24315.Element
-	point.SetRandom()
-	proof, err := kzg_bls24315.Open(f, point, srs.Pk)
-	assert.NoError(err)
+// 	// pick a number on byte shorter than the modulus size
+// 	var target big.Int
+// 	target.SetUint64(1)
+// 	nbBits := api.Compiler().Field().BitLen()
+// 	nn := ((nbBits+7)/8)*8 - 8
+// 	target.Lsh(&target, uint(nn))
 
-	if err = kzg_bls24315.Verify(&com, &proof, point, srs.Vk); err != nil {
-		t.Fatal("verify proof", err)
-	}
+// 	// create the wrapped hash function
+// 	whSnark, err := recursion.NewHash(api, &target, true)
+// 	if err != nil {
+// 		return err
+// 	}
 
-	wCmt, err := ValueOfCommitment[sw_bls24315.G1Affine](com)
-	assert.NoError(err)
-	wProof, err := ValueOfOpeningProof[sw_bls24315.Scalar, sw_bls24315.G1Affine](point, proof)
-	assert.NoError(err)
-	wVk, err := ValueOfVerifyingKey[sw_bls24315.G2Affine](srs.Vk)
-	assert.NoError(err)
+// 	res, err := verifier.deriveGamma(circuit.Point, circuit.Digests[:], circuit.ClaimedValues[:], whSnark)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	verifier.scalarApi.AssertIsEqual(&res, &circuit.Gamma)
+// 	return nil
+// }
 
-	assignment := KZGVerificationCircuit[sw_bls24315.Scalar, sw_bls24315.G1Affine, sw_bls24315.G2Affine, sw_bls24315.GT]{
-		VerifyingKey: wVk,
-		Commitment:   wCmt,
-		OpeningProof: wProof,
-	}
+// func TestDeriveGamma(t *testing.T) {
 
-	assert.CheckCircuit(&KZGVerificationCircuit[sw_bls24315.Scalar, sw_bls24315.G1Affine, sw_bls24315.G2Affine, sw_bls24315.GT]{}, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BW6_633))
-}
+// 	assert := test.NewAssert(t)
 
-func TestValueOfCommitment(t *testing.T) {
-	assert := test.NewAssert(t)
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bn254.Generators()
-		assignment, err := ValueOfCommitment[sw_bn254.G1Affine](G1)
-		assert.NoError(err)
-		_ = assignment
-	}, "bn254")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bls12377.Generators()
-		assignment, err := ValueOfCommitment[sw_bls12377.G1Affine](G1)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls12377")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bls12381.Generators()
-		assignment, err := ValueOfCommitment[sw_bls12381.G1Affine](G1)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls12381")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bw6761.Generators()
-		assignment, err := ValueOfCommitment[sw_bw6761.G1Affine](G1)
-		assert.NoError(err)
-		_ = assignment
-	}, "bw6761")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bls24315.Generators()
-		assignment, err := ValueOfCommitment[sw_bls24315.G1Affine](G1)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls24315")
-}
+// 	// prepare test data
+// 	alpha, err := rand.Int(rand.Reader, ecc.BN254.ScalarField())
+// 	assert.NoError(err)
+// 	srs, err := kzg_bn254.NewSRS(kzgSize, alpha)
+// 	assert.NoError(err)
 
-func TestValueOfOpeningProof(t *testing.T) {
-	assert := test.NewAssert(t)
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bn254.Generators()
-		var value, point fr_bn254.Element
-		value.SetRandom()
-		point.SetRandom()
-		proof := kzg_bn254.OpeningProof{
-			H:            G1,
-			ClaimedValue: value,
-		}
-		assignment, err := ValueOfOpeningProof[sw_bn254.Scalar, sw_bn254.G1Affine](point, proof)
-		assert.NoError(err)
-		_ = assignment
-	}, "bn254")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bls12377.Generators()
-		var value, point fr_bls12377.Element
-		value.SetRandom()
-		point.SetRandom()
-		proof := kzg_bls12377.OpeningProof{
-			H:            G1,
-			ClaimedValue: value,
-		}
-		assignment, err := ValueOfOpeningProof[sw_bls12377.Scalar, sw_bls12377.G1Affine](point, proof)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls12377")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bls12381.Generators()
-		var value, point fr_bls12381.Element
-		value.SetRandom()
-		point.SetRandom()
-		proof := kzg_bls12381.OpeningProof{
-			H:            G1,
-			ClaimedValue: value,
-		}
-		assignment, err := ValueOfOpeningProof[sw_bls12381.Scalar, sw_bls12381.G1Affine](point, proof)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls12381")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bw6761.Generators()
-		var value, point fr_bw6761.Element
-		value.SetRandom()
-		point.SetRandom()
-		proof := kzg_bw6761.OpeningProof{
-			H:            G1,
-			ClaimedValue: value,
-		}
-		assignment, err := ValueOfOpeningProof[sw_bw6761.Scalar, sw_bw6761.G1Affine](point, proof)
-		assert.NoError(err)
-		_ = assignment
-	}, "bw6761")
-	assert.Run(func(assert *test.Assert) {
-		_, _, G1, _ := bls24315.Generators()
-		var value, point fr_bls24315.Element
-		value.SetRandom()
-		point.SetRandom()
-		proof := kzg_bls24315.OpeningProof{
-			H:            G1,
-			ClaimedValue: value,
-		}
-		assignment, err := ValueOfOpeningProof[sw_bls24315.Scalar, sw_bls24315.G1Affine](point, proof)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls24315")
-}
+// 	var polynomials [10][]fr_bn254.Element
+// 	var digests [10]kzg_bn254.Digest
+// 	for i := 0; i < 10; i++ {
+// 		polynomials[i] = make([]fr_bn254.Element, polynomialSize)
+// 		for j := 0; j < polynomialSize; j++ {
+// 			polynomials[i][j].SetRandom()
+// 		}
+// 		digests[i], err = kzg_bn254.Commit(polynomials[i], srs.Pk)
+// 		assert.NoError(err)
+// 	}
 
-func TestValueOfSRS(t *testing.T) {
-	assert := test.NewAssert(t)
-	assert.Run(func(assert *test.Assert) {
-		_, _, _, G2 := bn254.Generators()
-		vk := kzg_bn254.VerifyingKey{
-			G2: [2]bn254.G2Affine{G2, G2},
-		}
-		assignment, err := ValueOfVerifyingKey[sw_bn254.G2Affine](vk)
-		assert.NoError(err)
-		_ = assignment
-	}, "bn254")
-	assert.Run(func(assert *test.Assert) {
-		_, _, _, G2 := bls12377.Generators()
-		vk := kzg_bls12377.VerifyingKey{
-			G2: [2]bls12377.G2Affine{G2, G2},
-		}
-		assignment, err := ValueOfVerifyingKey[sw_bls12377.G2Affine](vk)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls12377")
-	assert.Run(func(assert *test.Assert) {
-		_, _, _, G2 := bls12381.Generators()
-		vk := kzg_bls12381.VerifyingKey{
-			G2: [2]bls12381.G2Affine{G2, G2},
-		}
-		assignment, err := ValueOfVerifyingKey[sw_bls12381.G2Affine](vk)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls12381")
-	assert.Run(func(assert *test.Assert) {
-		_, _, _, G2 := bw6761.Generators()
-		vk := kzg_bw6761.VerifyingKey{
-			G2: [2]bw6761.G2Affine{G2, G2},
-		}
-		assignment, err := ValueOfVerifyingKey[sw_bw6761.G2Affine](vk)
-		assert.NoError(err)
-		_ = assignment
-	}, "bw6761")
-	assert.Run(func(assert *test.Assert) {
-		_, _, _, G2 := bls24315.Generators()
-		vk := kzg_bls24315.VerifyingKey{
-			G2: [2]bls24315.G2Affine{G2, G2},
-		}
-		assignment, err := ValueOfVerifyingKey[sw_bls24315.G2Affine](vk)
-		assert.NoError(err)
-		_ = assignment
-	}, "bls24315")
-}
+// 	var point fr_bn254.Element
+// 	point.SetRandom()
+// 	var target big.Int
+// 	target.SetUint64(1)
+// 	nbBits := ecc.BLS12_381.ScalarField().BitLen()
+// 	nn := ((nbBits+7)/8)*8 - 8
+// 	target.Lsh(&target, uint(nn))
+// 	h, err := recursion.NewShort(ecc.BLS12_381.ScalarField(), &target)
+// 	assert.NoError(err)
+
+// 	batchOpeningProof, err := kzg_bn254.BatchOpenSinglePoint(polynomials[:], digests[:], point, h, srs.Pk)
+// 	assert.NoError(err)
+
+// 	gamma, err := kzg_bn254.DeriveGamma(point, digests[:], batchOpeningProof.ClaimedValues, h)
+// 	assert.NoError(err)
+
+// 	// prepare witness
+// 	wPoint, err := ValueOfScalar[emulated.BN254Fr](point)
+// 	assert.NoError(err)
+// 	var wDigests [10]Commitment[sw_bn254.G1Affine]
+// 	for i := 0; i < 10; i++ {
+// 		wDigests[i], err = ValueOfCommitment[sw_bn254.G1Affine](digests[i])
+// 		assert.NoError(err)
+// 	}
+// 	var wClaimedValues [10]emulated.Element[emulated.BN254Fr]
+// 	for i := 0; i < 10; i++ {
+// 		wClaimedValues[i], err = ValueOfScalar[emulated.BN254Fr](batchOpeningProof.ClaimedValues[i])
+// 		assert.NoError(err)
+// 	}
+// 	wGmma, err := ValueOfScalar[emulated.BN254Fr](gamma)
+// 	assert.NoError(err)
+
+// 	assignment := DeriveGammaTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{
+// 		Point:         wPoint,
+// 		Digests:       wDigests,
+// 		ClaimedValues: wClaimedValues,
+// 		Gamma:         wGmma,
+// 	}
+
+// 	assert.CheckCircuit(&DeriveGammaTest[emulated.BN254Fr, sw_bn254.G1Affine, sw_bn254.G2Affine, sw_bn254.GTEl]{}, test.WithValidAssignment(&assignment), test.WithCurves(ecc.BLS12_381), test.WithBackends(backend.PLONK))
+// }
